refactor: 메인 캐러셀 컴포넌트 분리
- features/carousel/MainCarousel
- 페이지별 컴포넌트 분리
