+++ src/pages/browse/Movie/Movie.types.ts
@@ -1,0 +1,1 @@
+export type TMovieTab = 'popular' | 'now_playing' | 'top_rated' | 'upcoming';
