+++ src/pages/browse/Movie/Popular/Popular.styles.tsx
@@ -1,0 +1,5 @@
+import styled from '@emotion/styled';
+import { getFontStyle } from '@/styles/typography';
+import { COLORS } from '@/styles/colors';
+
+export const Container = styled.div``;
