+++ src/services/movie.types.ts
@@ -1,0 +1,10 @@
+export interface IPopularMovieResponse {
+  results: IPopularMovie[];
+}
+
+export interface IPopularMovie {
+  id: number;
+  backdrop_path: string;
+  title: string;
+  overview: string;
+}
